One error to go :-D Lexer = %80 Parser = %87 new expression and... that's all?
git-svn-id: http://svn.dsource.org/projects/descent/trunk@16 6538119e-2221-0410-ba12-9a5a7e5062c7

+++ descent.core/src/descent/core/dom/IIntegerExpression.java
@@ -1,5 +1,7 @@
 package descent.core.dom;
 
+import java.math.BigInteger;
+
 /**
  * An integer expression:
  */
@@ -8,6 +10,6 @@ public interface IIntegerExpression extends IExpression {
 	/**
 	 * Returns the integer.
 	 */
-	long getValue();
+	BigInteger getValue();
 
 }
